Add a Projects content-type
Is more trouble than you'd expect it to be. Need to add a 'collection'
field so that a post either belongs to 'blog' or 'projects'.

Create separate index pages for them both. Also need to separate them when
creating pages, so the pagination only belongs to a certain collection
(i.e. a blog post doesn't have a project post as a next-item)